Moved syscall def to arch specific
