+++ product-mini/platforms/linux/wali_arch/x86_64/signal/restore.s
